// apps/web/src/api/__generated__/types/UpdateCollectionRequest.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * sync
 * OpenAPI spec version: 0.0.1
 */

export interface UpdateCollectionRequest {
  /**
   * 변경할 컬렉션 이름 (생략 시 기존 값 유지, 최대 255자)
   * @nullable
   */
  name?: string | null;
  /**
   * 변경할 공개 여부 (생략 시 기존 값 유지)
   * @nullable
   */
  isPublic?: boolean | null;
  /**
   * 변경할 컬렉션 설명 (생략 시 기존 값 유지)
   * @nullable
   */
  description?: string | null;
}